move user value objects to slice
